+++ src/models/index.ts
@@ -2,14 +2,17 @@ import * as coreClient from "@azure/core-client";
 import * as coreRestPipeline from "@azure/core-rest-pipeline";
 
 export interface GetAllDocumentsResults {
+  /** Number of documents in result */
   count?: number;
+  /** URL to request next page of results */
   next?: string;
+  /** URL to request previous page of results */
   previous?: string;
   results?: Meta[];
 }
 
 export interface Meta {
-  /** Unique identifier for the resume. If creating a document and left blank, one will be automatically generated. */
+  /** Unique identifier for the document. If creating a document and left blank, one will be automatically generated. */
   identifier: string | null;
   /** Optional filename of the file */
   fileName?: string;
@@ -19,7 +22,7 @@ export interface Meta {
   readyDt?: Date;
   /** If true, some exception was raised during processing. Check the 'error' field of the main return object. */
   failed: boolean;
-  /** The date/time in ISO-8601 format when the resume will be automatically deleted.  Defaults to no expiry. */
+  /** The date/time in ISO-8601 format when the document will be automatically deleted.  Defaults to no expiry. */
   expiryTime?: string;
 }
 
@@ -58,7 +61,7 @@ export interface ResumeData {
   publications?: string[];
   referees?: ResumeDataRefereesItem[];
   sections?: ResumeDataSectionsItem[];
-  /** Probability that the given document is a resume. Values below 30 suggest that the resume is not a resume. */
+  /** Probability that the given document is a resume. Values below 30 suggest that the document is not a resume. */
   isResumeProbability?: number;
   /** All of the raw text of the parsed resume, example is shortened for readiblity */
   rawText?: string;
@@ -165,20 +168,24 @@ export interface RedactedResume {
 }
 
 export interface RedactedResumeData {
+  /** URL to redacted PDF */
   redactedPdf?: string;
 }
 
 export interface Paths1UtuacyResumeFormatsGetResponses200ContentApplicationJsonSchema {
+  /** Number of documents in result */
   count?: number;
+  /** URL to request next page of results */
   next?: string;
+  /** URL to request previous page of results */
   previous?: string;
   results?: Get200ApplicationJsonPropertiesItemsItem[];
 }
 
 export interface Get200ApplicationJsonPropertiesItemsItem {
-  /** Unique identifier for the resume. If creating a document and left blank, one will be automatically generated. */
+  /** Unique identifier for the document. If creating a document and left blank, one will be automatically generated. */
   identifier: string | null;
-  /** The template to apply */
+  /** URL to a template to apply */
   formatFile: string;
 }
 
@@ -192,10 +199,64 @@ export interface ReformattedResumeData {
   reformattedFile?: string;
 }
 
+export interface GetAllInvoicesResults {
+  /** Number of documents in result */
+  count?: number;
+  /** URL to request next page of results */
+  next?: string;
+  /** URL to request previous page of results */
+  previous?: string;
+  results?: Meta[];
+}
+
+export interface Invoice {
+  data: InvoiceData | null;
+  meta: Meta;
+  error: ErrorModel;
+}
+
+export interface InvoiceData {
+  invoiceDate?: string;
+  invoiceOrderDate?: string;
+  paymentDateDue?: string;
+  paymentAmountBase?: string;
+  paymentAmountTax?: string;
+  paymentAmountTotal?: string;
+  paymentAmountPaid?: string;
+  paymentAmountDue?: string;
+  invoiceNumber?: string;
+  invoicePurchaseOrderNumber?: string;
+  supplierBusinessNumber?: string;
+  customerNumber?: string;
+  customerBusinessNumber?: string;
+  paymentReference?: string;
+  bankAccountNumber?: string;
+  supplierVAT?: string;
+  customerVAT?: string;
+  bpayBillerCode?: string;
+  bpayReference?: string;
+  bankSortCode?: string;
+  bankIban?: string;
+  bankSwift?: string;
+  bankBSB?: string;
+  customerContactName?: string;
+  customerCompanyName?: string;
+  supplierCompanyName?: string;
+  customerBillingAddress?: Location;
+  customerDeliveryAddress?: Location;
+  supplierAddress?: Location;
+  customerPhoneNumber?: string;
+  supplierPhoneNumber?: string;
+  supplierFax?: string;
+  customerEmail?: string;
+  supplierEmail?: string;
+  supplierWebsite?: string;
+}
+
 export interface Paths7EskthResumesPostRequestbodyContentMultipartFormDataSchema {
   /** File as binary data blob */
   file?: coreRestPipeline.RequestBodyType;
-  /** Unique identifier for the resume. If creating a document and left blank, one will be automatically generated. */
+  /** Unique identifier for the document. If creating a document and left blank, one will be automatically generated. */
   identifier?: string;
   /** Optional filename of the file */
   fileName?: string;
@@ -204,22 +265,22 @@ export interface Paths7EskthResumesPostRequestbodyContentMultipartFormDataSchema
   /** If "true" (default), will return a response only after processing has completed. If "false", will return an empty data object which can be polled at the GET endpoint until processing is complete. */
   wait?: string;
   /** Language code in ISO 639-1 format. Must specify zh-cn or zh-tw for Chinese. */
-  resumeLanguage?: string;
-  /** The date/time in ISO-8601 format when the resume will be automatically deleted.  Defaults to no expiry. */
+  language?: string;
+  /** The date/time in ISO-8601 format when the document will be automatically deleted.  Defaults to no expiry. */
   expiryTime?: string;
 }
 
 export interface Paths8DdhfcRedactedResumesPostRequestbodyContentMultipartFormDataSchema {
   /** File as binary data blob */
   file?: coreRestPipeline.RequestBodyType;
-  /** Unique identifier for the resume. If creating a document and left blank, one will be automatically generated. */
+  /** Unique identifier for the document. If creating a document and left blank, one will be automatically generated. */
   identifier?: string;
   /** Optional filename of the file */
   fileName?: string;
   /** URL to file to download and process */
   url?: string;
   /** Language code in ISO 639-1 format. Must specify zh-cn or zh-tw for Chinese. */
-  resumeLanguage?: string;
+  language?: string;
   /** If "true" (default), will return a response only after processing has completed. If "false", will return an empty data object which can be polled at the GET endpoint until processing is complete. */
   wait?: string;
   /** Whether to redact headshot */
@@ -236,25 +297,42 @@ export interface Paths8DdhfcRedactedResumesPostRequestbodyContentMultipartFormDa
   redactLocations?: string;
   /** Whether to redact dates */
   redactDates?: string;
-  /** The date/time in ISO-8601 format when the resume will be automatically deleted.  Defaults to no expiry. */
+  /** The date/time in ISO-8601 format when the document will be automatically deleted.  Defaults to no expiry. */
   expiryTime?: string;
 }
 
 export interface PathsYzn84IReformattedResumesPostRequestbodyContentMultipartFormDataSchema {
   /** File as binary data blob */
   file?: coreRestPipeline.RequestBodyType;
-  /** Unique identifier for the resume. If creating a document and left blank, one will be automatically generated. */
+  /** Unique identifier for the document. If creating a document and left blank, one will be automatically generated. */
   identifier?: string;
   /** Optional filename of the file */
   fileName?: string;
   /** URL to file to download and process */
   url?: string;
   /** Language code in ISO 639-1 format. Must specify zh-cn or zh-tw for Chinese. */
-  resumeLanguage?: string;
+  language?: string;
   /** Identifier of the format used */
   resumeFormat: string;
   /** If "true" (default), will return a response only after processing has completed. If "false", will return an empty data object which can be polled at the GET endpoint until processing is complete. */
   wait?: string;
+}
+
+export interface Paths1BwrvmkInvoicesPostRequestbodyContentMultipartFormDataSchema {
+  /** File as binary data blob */
+  file?: coreRestPipeline.RequestBodyType;
+  /** Unique identifier for the document. If creating a document and left blank, one will be automatically generated. */
+  identifier?: string;
+  /** Optional filename of the file */
+  fileName?: string;
+  /** URL to file to download and process */
+  url?: string;
+  /** If "true" (default), will return a response only after processing has completed. If "false", will return an empty data object which can be polled at the GET endpoint until processing is complete. */
+  wait?: string;
+  /** Language code in ISO 639-1 format. Must specify zh-cn or zh-tw for Chinese. */
+  language?: string;
+  /** The date/time in ISO-8601 format when the document will be automatically deleted.  Defaults to no expiry. */
+  expiryTime?: string;
 }
 
 /** Optional parameters. */
@@ -269,7 +347,7 @@ export interface AffindaAPICreateResumeOptionalParams
   extends coreClient.OperationOptions {
   /** File as binary data blob */
   file?: coreRestPipeline.RequestBodyType;
-  /** Unique identifier for the resume. If creating a document and left blank, one will be automatically generated. */
+  /** Unique identifier for the document. If creating a document and left blank, one will be automatically generated. */
   identifier?: string;
   /** Optional filename of the file */
   fileName?: string;
@@ -278,8 +356,8 @@ export interface AffindaAPICreateResumeOptionalParams
   /** If "true" (default), will return a response only after processing has completed. If "false", will return an empty data object which can be polled at the GET endpoint until processing is complete. */
   wait?: string;
   /** Language code in ISO 639-1 format. Must specify zh-cn or zh-tw for Chinese. */
-  resumeLanguage?: string;
-  /** The date/time in ISO-8601 format when the resume will be automatically deleted.  Defaults to no expiry. */
+  language?: string;
+  /** The date/time in ISO-8601 format when the document will be automatically deleted.  Defaults to no expiry. */
   expiryTime?: string;
 }
 
@@ -312,7 +390,7 @@ export interface AffindaAPICreateRedactedResumeOptionalParams
   extends coreClient.OperationOptions {
   /** File as binary data blob */
   file?: coreRestPipeline.RequestBodyType;
-  /** Unique identifier for the resume. If creating a document and left blank, one will be automatically generated. */
+  /** Unique identifier for the document. If creating a document and left blank, one will be automatically generated. */
   identifier?: string;
   /** Optional filename of the file */
   fileName?: string;
@@ -321,8 +399,8 @@ export interface AffindaAPICreateRedactedResumeOptionalParams
   /** If "true" (default), will return a response only after processing has completed. If "false", will return an empty data object which can be polled at the GET endpoint until processing is complete. */
   wait?: string;
   /** Language code in ISO 639-1 format. Must specify zh-cn or zh-tw for Chinese. */
-  resumeLanguage?: string;
-  /** The date/time in ISO-8601 format when the resume will be automatically deleted.  Defaults to no expiry. */
+  language?: string;
+  /** The date/time in ISO-8601 format when the document will be automatically deleted.  Defaults to no expiry. */
   expiryTime?: string;
   /** Whether to redact headshot */
   redactHeadshot?: string;
@@ -376,7 +454,7 @@ export interface AffindaAPICreateReformattedResumeOptionalParams
   extends coreClient.OperationOptions {
   /** File as binary data blob */
   file?: coreRestPipeline.RequestBodyType;
-  /** Unique identifier for the resume. If creating a document and left blank, one will be automatically generated. */
+  /** Unique identifier for the document. If creating a document and left blank, one will be automatically generated. */
   identifier?: string;
   /** Optional filename of the file */
   fileName?: string;
@@ -385,7 +463,7 @@ export interface AffindaAPICreateReformattedResumeOptionalParams
   /** If "true" (default), will return a response only after processing has completed. If "false", will return an empty data object which can be polled at the GET endpoint until processing is complete. */
   wait?: string;
   /** Language code in ISO 639-1 format. Must specify zh-cn or zh-tw for Chinese. */
-  resumeLanguage?: string;
+  language?: string;
 }
 
 /** Contains response data for the createReformattedResume operation. */
@@ -404,6 +482,49 @@ export interface AffindaAPIDeleteReformattedResumeOptionalParams
 
 /** Contains response data for the deleteReformattedResume operation. */
 export type AffindaAPIDeleteReformattedResumeResponse = RequestError;
+
+/** Optional parameters. */
+export interface AffindaAPIGetAllInvoicesOptionalParams
+  extends coreClient.OperationOptions {}
+
+/** Contains response data for the getAllInvoices operation. */
+export type AffindaAPIGetAllInvoicesResponse = GetAllInvoicesResults;
+
+/** Optional parameters. */
+export interface AffindaAPICreateInvoiceOptionalParams
+  extends coreClient.OperationOptions {
+  /** File as binary data blob */
+  file?: coreRestPipeline.RequestBodyType;
+  /** Unique identifier for the document. If creating a document and left blank, one will be automatically generated. */
+  identifier?: string;
+  /** Optional filename of the file */
+  fileName?: string;
+  /** URL to file to download and process */
+  url?: string;
+  /** If "true" (default), will return a response only after processing has completed. If "false", will return an empty data object which can be polled at the GET endpoint until processing is complete. */
+  wait?: string;
+  /** Language code in ISO 639-1 format. Must specify zh-cn or zh-tw for Chinese. */
+  language?: string;
+  /** The date/time in ISO-8601 format when the document will be automatically deleted.  Defaults to no expiry. */
+  expiryTime?: string;
+}
+
+/** Contains response data for the createInvoice operation. */
+export type AffindaAPICreateInvoiceResponse = Invoice;
+
+/** Optional parameters. */
+export interface AffindaAPIGetInvoiceOptionalParams
+  extends coreClient.OperationOptions {}
+
+/** Contains response data for the getInvoice operation. */
+export type AffindaAPIGetInvoiceResponse = Invoice;
+
+/** Optional parameters. */
+export interface AffindaAPIDeleteInvoiceOptionalParams
+  extends coreClient.OperationOptions {}
+
+/** Contains response data for the deleteInvoice operation. */
+export type AffindaAPIDeleteInvoiceResponse = RequestError;
 
 /** Optional parameters. */
 export interface AffindaAPIOptionalParams