+++ CHANGELOG.md
@@ -5,9 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Invoices endpoint
+
+### Removed
+- Removed 'url' format from url strings in api spec
 
 ## [0.2.5] - 2021-09-30
-
 ### Added
 - Adding LinkedIn to ResumeData